+++ sql/queries/ratings.sql
@@ -1,8 +1,8 @@
 -- name: InsertRating :one
 INSERT INTO ratings (
-    id, title_id, user_id, note, created_at, updated_at
+    id, title_id, user_id, group_id, note, created_at, updated_at
 ) VALUES (
-    $1, $2, $3, $4, $5, $6
+    $1, $2, $3, $4, $5, $6, $7
 )
 RETURNING *;
 
@@ -16,13 +16,13 @@ RETURNING *;
 SELECT * FROM ratings WHERE id = $1 AND user_id = $2;
 
 -- name: GetRatingRowByUserTitle :one
-SELECT * FROM ratings WHERE user_id = $1 AND title_id = $2;
+SELECT * FROM ratings WHERE user_id = $1 AND title_id = $2 AND group_id = $3;
 
 -- name: GetRatingRowsByTitleId :many
-SELECT * FROM ratings WHERE title_id = $1;
+SELECT * FROM ratings WHERE title_id = $1 AND group_id = $2;
 
 -- name: GetRatingRowsByTitleIds :many
-SELECT * FROM ratings WHERE title_id = ANY($1::text[]);
+SELECT * FROM ratings WHERE title_id = ANY($1::text[]) AND group_id = $2;
 
 -- name: DeleteRatingRow :execrows
 DELETE FROM ratings WHERE id = $1 AND user_id = $2;